Improve Category Error Handling closes #19

+++ backend/src/main/java/com/domaciproizvodi/service/OrderItemService.java
@@ -1,9 +1,8 @@
 package com.domaciproizvodi.service;
 
+import com.domaciproizvodi.exceptions.OrderItemNotFOundException;
 import com.domaciproizvodi.model.OrderItem;
 import com.domaciproizvodi.repository.OrderItemRepository;
-import com.domaciproizvodi.repository.OrderRepository;
-import com.domaciproizvodi.repository.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -16,12 +15,6 @@ public class OrderItemService {
     @Autowired
     private OrderItemRepository orderItemRepository;
 
-    @Autowired
-    private OrderRepository orderRepository;
-
-    @Autowired
-    private ProductRepository productRepository;
-
     public List<OrderItem> getAllOrderItems() {
         return orderItemRepository.findAll();
     }
@@ -31,8 +24,6 @@ public class OrderItemService {
     }
 
     public OrderItem createOrderItem(OrderItem orderItem) {
-
-
         return orderItemRepository.save(orderItem);
     }
 
@@ -45,10 +36,13 @@ public class OrderItemService {
                     orderItem.setOrder(updatedOrderItem.getOrder());
                     return orderItemRepository.save(orderItem);
                 })
-                .orElseThrow(() -> new RuntimeException("Order item not found"));
+                .orElseThrow(() -> new OrderItemNotFOundException("Order item not found with id: " + id));
     }
 
     public void deleteOrderItem(Long id) {
+        if (!orderItemRepository.existsById(id)) {
+            throw new OrderItemNotFOundException("Order item not found with id: " + id);
+        }
         orderItemRepository.deleteById(id);
     }
